신청취소시 예외처리 추가(DB에 없는 신청자가 취소 요청을 해도 취소완료 메세지)

+++ src/main/java/com/example/miniproject14/service/ApplicantsService.java
@@ -12,6 +12,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class ApplicantsService {
@@ -38,12 +40,23 @@ public class ApplicantsService {
         if(user == null){
             throw new IllegalArgumentException("로그인이 필요합니다");
         }
-        try {
-            applicantsRepository.deleteByUserIdAndBoardId(user.getId(), applicantsRequestDto.getBoardId());
-            return new StatusResponseDto("신청이 취소되었습니다.", HttpStatus.OK); // DB에 정상적으로 저장 되었을 경우 결과 리턴
-        }catch(Exception e){
-            return new StatusResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST);
+        Optional<Applicants> existingApplicant = applicantsRepository.findByUserIdAndBoardId(user.getId(), applicantsRequestDto.getBoardId());
+        if (existingApplicant.isPresent()) {
+            try {
+                applicantsRepository.deleteByUserIdAndBoardId(user.getId(), applicantsRequestDto.getBoardId());
+                return new StatusResponseDto("신청이 취소되었습니다.", HttpStatus.OK);
+            } catch(Exception e){
+                return new StatusResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST);
+            }
+        } else {
+            return new StatusResponseDto("신청하지 않은 게시물입니다.", HttpStatus.BAD_REQUEST);
         }
+//        try {
+//            applicantsRepository.deleteByUserIdAndBoardId(user.getId(), applicantsRequestDto.getBoardId());
+//            return new StatusResponseDto("신청이 취소되었습니다.", HttpStatus.OK); // DB에 정상적으로 저장 되었을 경우 결과 리턴
+//        }catch(Exception e){
+//            return new StatusResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST);
+//        }
     }
